rag-a-muffin/rag_fast_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
pytest
pytest-asyncio
httpx
weaviate-client==4.16.9
requests>=2.31.0